set uid to cookie
