+++ utils/apollo-client.ts
@@ -1,6 +1,6 @@
 import { ApolloClient, createHttpLink, InMemoryCache } from "@apollo/client";
 import { setContext } from "@apollo/client/link/context";
-import { tokenKeyName } from "./token-key-name";
+import { tokenKeyName } from "./cookie-key-names";
 
 const httpLink = createHttpLink({
   uri: `${process.env.NEXT_PUBLIC_BACKEND_URL}/graphql`,
