+++ utils/cookie-key-names.ts
@@ -1,1 +1,2 @@
 export const tokenKeyName = "token";
+export const uidKeyName = "uid";
